+++ RedPandaIDE/settingsdialog/environmentappearencewidget.cpp
@@ -97,6 +97,7 @@ void EnvironmentAppearenceWidget::init()
     ui->cbLanguage->addItem(tr("English"),"en");
     ui->cbLanguage->addItem(tr("Portuguese"),"pt");
     ui->cbLanguage->addItem(tr("Simplified Chinese"),"zh_CN");
+    ui->cbLanguage->addItem(tr("Traditional Chinese"),"zh_TW");
     QList<PIconSet> iconSets = pIconsManager->listIconSets();
     foreach(const PIconSet& iconSet, iconSets) {
         ui->cbIconSet->addItem(iconSet->displayName,iconSet->name);